# docs/api/openapi/mesas.openapi.yaml
...
openapi: "3.0.3"
info:
  title: "Mesas API — Artifício RPG"
  description: "API do módulo Mesas. Metadados x-artificio-* definidos por heurística de path."
  version: "0.1.0"
servers:
  - url: https://mesas.artificiorpg.com
    description: Produção
  - url: http://localhost:4000
    description: Desenvolvimento
security: []
paths:
...
  /api/v1/gm/me:
    get:
      operationId: getapi_v1_gm_me
      summary: "Consulta api v1 gm me"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/me:
    get:
      operationId: getapi_v1_me
      summary: "Consulta api v1 me"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/profile/gm:
    patch:
      operationId: patchapi_v1_profile_gm
      summary: "Atualiza api v1 profile gm"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/profile/me:
    get:
      operationId: getapi_v1_profile_me
      summary: "Consulta api v1 profile me"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
    patch:
      operationId: patchapi_v1_profile_me
      summary: "Atualiza api v1 profile me"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/me/connect/discord:
    delete:
      operationId: deleteapi_v1_profile_me_connect_discord
      summary: "Remove api v1 profile me connect discord"
      x-artificio-owner: mesas
      x-artificio-scope: self-service
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
    post:
      operationId: postapi_v1_profile_me_connect_discord
      summary: "Cria ou executa api v1 profile me connect discord"
      x-artificio-owner: mesas
      x-artificio-scope: self-service
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/me/discord:
    get:
      operationId: getapi_v1_profile_me_discord
      summary: "Consulta api v1 profile me discord"
      x-artificio-owner: mesas
      x-artificio-scope: self-service
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/me/gm:
    patch:
      operationId: patchapi_v1_profile_me_gm
      summary: "Atualiza api v1 profile me gm"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/me/google-picture:
    post:
      operationId: postapi_v1_profile_me_google_picture
      summary: "Cria ou executa api v1 profile me google-picture"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/me/player:
    patch:
      operationId: patchapi_v1_profile_me_player
      summary: "Atualiza api v1 profile me player"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/me/profile:
    patch:
      operationId: patchapi_v1_profile_me_profile
      summary: "Atualiza api v1 profile me profile"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/player:
    patch:
      operationId: patchapi_v1_profile_player
      summary: "Atualiza api v1 profile player"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/systems:
    post:
      operationId: postapi_v1_profile_systems
      summary: "Cria ou executa api v1 profile systems"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/profile/systems/{id}:
    delete:
      operationId: deleteapi_v1_profile_systems_id
      summary: "Remove api v1 profile systems id"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/systems:
    get:
      operationId: getapi_v1_systems
      summary: "Consulta api v1 systems"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: none
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...